removed Calendar redux store
