+++ frontend/src/routes/FacilityBooking/CreateBookingDailyView/index.tsx
@@ -71,7 +71,7 @@ export default function CreateBookingDailyView() {
   const { selectedFacilityName, isLoading, selectedStartTime } = useSelector(
     (state: RootState) => state.facilityBooking,
   )
-  const { clickedDate } = useSelector((state: RootState) => state.calendar)
+  const { clickedDate } = useSelector((state: RootState) => state.facilityBooking)
 
   const selectedFacilityId = parseInt(params.facilityId)
